cmds: simplify ApplyDiffSelection
We do not need to re-apply the diff when applying a diff selection as it
happens automatically when the view restores its selection.

Update the index so that we do not need to depend on the "git diff"
side-effect of updating the index, without which we would not be
detecting when files change state from staged <=> unstaged.

Signed-off-by: David Aguilar <davvid@gmail.com>