core: move unicode handling to the outer edges
There were too many instances of encode() and decode() throughout the
code.  Centralize everything into the `core` module so that we can keep
the internals unicode-pure.

This simplifies things substantially since unicode is now used
consistently throughout and the only places where encode() and decode()
are used are in the outer edges and a few spot edge cases.  The majority
of the code is now blissfully ignorant.

Signed-off-by: David Aguilar <davvid@gmail.com>