git: always return (status, out, err)
Git.execute() had too many flags and its output format changed depending
on these flags.

Simplify it by making git.foo() always return (status, out, err).  This
makes the code easier to read and more robust since we will never try to
parse stderr as if it were stdout.

This is a fairly big patch since it touched pretty much every file that
contained git commands.  We may want to later take some things (e.g. the
stuff in stash.model) and move it to gitcmds so that later refactors are
easier and involve less files.

For example, eliminating global git and model instances would be helped
by such a change, and is a refactoring that needs to be done to allow
hot-swapping of the current Git repository, which would be a nice
feature.

Signed-off-by: David Aguilar <davvid@gmail.com>