grep: avoid searching for empty strings
Rename 'txt' to 'query' for consistency.  Avoid 'query' being empty by
checking its value in search(), which allows us to do without the extra
input_txt_changed() function.

Previously, switching modes would have triggered a search for the "hint"
text because we were using input_txt.as_unicode() instead of
input_txt.value(), which was a bug.  Fixing this causes the widget to
search for the empty string, which is avoided by checking for len(query)
in search() itself.

Signed-off-by: David Aguilar <davvid@gmail.com>