models.main: exit amend mode when merging
Detect when we've somehow a state where we're amending and in the middle
of a merge.  Reset the mode back to "none" when this happens.

This can happen if the user enters "amend" mode and then merges a
branch, either through the UI or via the command-line.  We now avoid
this undesirable state by automatically disabling the "amend" mode when
an update detects that we are in the middle of a merge.

Signed-off-by: David Aguilar <davvid@gmail.com>