commitmsg: better copy/paste behavior
Qt allows users to paste newlines into the `summary` field even though
it's not something that's normally allowed via the keyboard.  This is
unwanted because the commit message may contain more than what is
visible in the single-line field.

Allow users to paste multiple lines of text into the `summary` field by
automatically splitting it apart so that everything beyond the first
newline is moved to the `Extended description` field.

This prevents newlines from entering the `summary` field and improves
the editor's usability when pasting multiple lines of text.

Closes #212

Reported-by: @grandinj via github.com
Signed-off-by: David Aguilar <davvid@gmail.com>