cmds: do not capture stdout and stderr when rebasing
Git prints a lot of information to the shell when rebasing.  Set stdout
and stderr to None so that Python passes stdout and stderr through to
the calling shell.  This makes it so that the in-progress prints from
`git rebase` is seen when doing e.g. `git cola rebase origin/master`.

Signed-off-by: David Aguilar <davvid@gmail.com>