cmds: adjust how we capture stdout/stderr when rebasing
3c7d62fb72eaac6ccb681ef1d61dd5f82b5c9845 modified how we launch
"git rebase" so that we do not capture stdout and stderr.

This is really only useful when git-xbase is launched from the shell via
"git cola rebase".  We do not want anything going to the console when
launched from the main UI..

Modify the Rebase command so that it captures its output by when
invoked from the main UI.  Do not capture output when invoked
via "git cola rebase".

Signed-off-by: David Aguilar <davvid@gmail.com>