commitmsg: remove event filter to avoid spurious resizing
The commit message editor changes its minimum size when its focus
changes.  This allowed a nice behavior where it can be resized down so
that the description is not visible.

This is problematic because changing focus to/from the commit message
editor causes the dock widgets to resize themselves, which is not
wanted.

Set a sane minimum size on the description field and remove the event
filters so that we no longer futz with the minimum size.  This prevents
the spurious resizing.

Signed-off-by: David Aguilar <davvid@gmail.com>