dag: Add modified files list to dag
Currently the diff viewer in dag lists all
changes within a commit. There is no way to
refine the diff to a single file. This is
problematic for a commit which have many files
modified.

Add a file widget to dag. It will display the
list of modified files when a commit selected.
Selecting a file in the list will make the diff
viewer display the diff for that file only.

Suggested-By: @systemovich via github.com in #217

Signed-off-by: Minarto Margoliono
               <lie.r.min.g@gmail.com>