dag: remove references to self.Mixin
The old bind_mixin() setup in cola.widgets.standard provided self.Mixin
as a way to refer to the widget-specific mixin class.

This is no longer needed since we can simply refer to the base
class instead.  In hindsight, we probably could have done without
this all along, but it was written the long-hand way before to avoid
any confusion about which method is being called.

Signed-off-by: David Aguilar <davvid@gmail.com>