Fixed build-git-cola.sh.
$install_dir does not exist. Instead, changed it to $COLA_PREFIX.

$COMPILE_TOP does not eixst. Instead, changed it to $compile_top.

The command to make git-cola was incorrect as it uses the currently
installed Python. Instead, it should be using the Python that was just
built and installed. Modified the make command to include overriding the
PYTHON value.

After fixing the make command, the extra script that adds environment
variables to run git-cola are no longer needed. Removed the notes about
the script and added notes to simply use git-cola from the install.