widgets: Fix faulty save_state() under Python 3
The result of `toBase64().data()`, bytes data, was wrapped into a string
using `ustr()` causing the resulting trying to be a *representation* of
the bytes (wrapped into `b''`), thus being buggy data.

I've replaced this with a proper ".decode()" call.

I've tried keeping it as bytes, but that didn't go through
serialization.

Fixes #236

Signed-off-by: Virgil Dupras <hsoft@hardcoded.net>