Add support for session management of window managers
Window managers can restart applications that were running when
the user logged off. This commit saves the current repository path
to the settings file if window manager requests the saving of the state.

On next startup the saved repository is loaded (if git-cola is started
in a directory without a repository).