zh_TW.po: update translation of "Help"
Currently, the menu named "Help" is directly translated as "幫助"(may read
as verb/noun in Chinese).  I "patched" it to "尋求幫助(Seeking for help)"
hoping it to be more sensible for users who are really need to find some
more help on using git-cola.

This "patch" is indeed "this particular translator's own preference and
convention" rather than "convention of general zh_TW localization
community", so ignore it if you want to keep safe :-).

Signed-off-by: Ｖ字龍(Vdragon) <pika1021@gmail.com>