widgets.action: provide tooltips for action buttons
Qt may truncate the button labels so provide a tooltip with the original
string.  This ensures that users can always have a way to see what a
button does even if there is not enough space on the screen to display
the label.

A nicer long-term solution may be to have a unique icon for each action
so that the tooltip is not "redundant", but finding an expressive
pictogram for each action is not easy.

Closes #292

Reported-by: Ｖ字龍(Vdragon) <pika1021@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>