dag: rework layout to avoid sizing issues
We were adding to the bottom dock area, which Qt implicitly keeps sized
to a large area that cannot be resized.  This means that the default
layout used by git-dag creates a diff widget that cannot be made
smaller.

Use the left and right regions instead so that each widget can be freely
sized.

Closes #299

Reported-by: Stanisław Halik <sthalik@misaki.pl>
Signed-off-by: David Aguilar <davvid@gmail.com>