widgets.diff: better selection action enabling
In the DiffEditor, the actions for acting upon the selected lines were
initially enabled, which meant that they could be selected in the
context menu if no text selection had ever been made, in which case they
would operate on the entire hunk under the cursor.  Once the first text
selection had been made, they would be enabled and disabled properly.
Fix this by disabling them when they are created.

Additionally, the action_process_selection action was always enabled,
which meant that if the user pressed the "S" shortcut key in the diff
pane when no text was selected, the action would be applied to the
entire hunk containing the cursor.  Fix this by enabling and disabling
this action in the same way as the other actions that operate on the
selected lines.

Signed-off-by: Daniel Harding <dharding@living180.net>