test.helper: shutil.rmtree workaround for Windows
On Windows, shutil.rmtree() raises an exception if it tries to remove a
file marked as read-only.  The files that git creates in the
.git/objects directory are read-only, so the GitRepositoryTestCase hits
this problem when shutil.rmtree() is called during tearDown.  Work
around the issue by passing a an onerror handler to shutil.rmtree()
which detects the case of trying to remove a read-only file, changes the
file to be writeable, and tries again to remove it.

Signed-off-by: Daniel Harding <dharding@living180.net>