test: make tests portable to Windows
Using os.system() to shell out to perform various test setup functions
does not work well on Windows because it uses different shell syntax.
To address this, remove the shell() method from TmpPathTestCase, along
with the shell() and pipe() functions from the test.helper module.  To
facilitate replacing usages of the shell() method, add three static
methods to TmpPathTestCase:

 * touch(): creates one or more empty files if they do not already
   exist.
 * write_file(): sets the content of file at the specified path
 * append_file(): appends content to the file at the specified path

In addition, add a git() method to GitRepositoryTestCase, which uses
subprocess to invoke git with the specified arguments and returns any
output generated by git.

Re-implement all the uses of the shell() method or the pipe() function
using these helper methods or built-in Python functions.

Signed-off-by: Daniel Harding <dharding@living180.net>