remote: add an explicit argument for the window title
We were combining the window title and the text that we use to track
which action to take, which is complicated and makes the usage of N_()
un-idiomatic since we end up using it on a variable instead of a string.

Pass the window title separately from the desired action to make things
simpler.

Signed-off-by: David Aguilar <davvid@gmail.com>