Status: Add autocompletion for the filter
To use the filter in status widget, a user need to know the exact path
they want to filter. This is a bit user unfriendly as the user might not
always remember the exact path they want to filter on.

This change will add autocomplete pop-up for filter in status widget. It
will autocomplete to both folders path and files path.

Reported-by: hrj
Signed-off-by: Minarto Margoliono <lie.r.min.g@gmail.com>