git-cola: add a --help-commands option
We previously special-cased '-h' and '--help' in sys.argv so that we
could print the available sub-commands help message when running
`git cola --help`.

It makes more sense to print the cola command's help message since it is
the default command that is when no sub-command is specified.

Add a separate '--help-commands' option that can be used to print the
available sub-commands.

Signed-off-by: David Aguilar <davvid@gmail.com>