commitmsg: improve summary down-key behavior
2e2b8937e7a47bf36d186f80f9ffa3c9068bd71d simplified how we handle the
focus switching between the summary and description fields but it lost
the ability to quickly jump to the end of the summary line by hitting
the "down" arrow.

Detect when the cursor is in the middle of the summary field and move
the cursor to the end of the line when "down" is pressed.  Give the
description field focus when the cursor is already at the end of the
line.

Signed-off-by: David Aguilar <davvid@gmail.com>