widgets.completion: simplify completion handling
Simplify the completion code by doing less.

We now do everything in a single KeyReleaseEvent() where we special-case
Tab, Enter et. al  and unconditionally do completion otherwise.

Ensure that the completion popup always shows the first item.  This
makes it so that pressing the down arrow will always highlight the first
visible entry.  The old code would allow the popup's scroll area to
bounce around, which resulted in the visible completions not matching
what is selected when pressing the down arrow.

Remove the mouse events since they complicate things and mess with the
standard interactions.  The callbacks were there to make the completion
popup appear when clicking in the widget.  This was helpful, but it also
made it hard to move the cursor when editing the line.

The moment the user presses a key (e.g. an arrow key) then the
completions will appear so the mouse events can be removed without
losing any functionality.

The end result is simpler code and simpler user interaction.

Signed-off-by: David Aguilar <davvid@gmail.com>