widgets.completion: do not offer completions when text has no matches
The "show all refs when no matches are found" makes the filename matches
found by the GitLogCompletionModel hard to see since they are buried
below the list of refs.

Remove this feature so that the actual matches bubble to the top rather
than being obscured by a potentially huge list of refs.

Signed-off-by: David Aguilar <davvid@gmail.com>