gui: update branch name in title after renaming a branch
After renaming current branch, the UI is never notified of the model
change.

Update the current branch reference after renaming a branch. This is
done even when the renamed branch is not currently active one. The
operation should not be expensive, and doing so will simplify the code
logic.

fix: #369
Reported-by: Ｖ字龍(Vdragon) <Vdragon.Taiwan@gmail.com>
Signed-off-by: Minarto Margoliono <lie.r.min.g@gmail.com>