dag:Add context menu for showing file history
Currently there is no easy way to track what recently change for a
particular file. This is useful to check which recent commit(s) actually
introduce the bug.

This change adds a context menu 'Show History' for filelist widget item
in DAG. When selected, the DAG will filter the log to display commits
which modified the selected file.

Signed-off-by: Minarto Margoliono <lie.r.min.g@gmail.com>