widgets/search: fix for Python 3
The Search dialog did not work under Python 3.  This is due to the fact
that the result of calling the builtin function map() was being passed
to the QListWidget.addItems() method in the Search.set_commit_list()
method.  This worked fine under Python 2 because map() returns a list.
Under Python 3 however it returns an iterator, which causes PyQt to
raise an exception.  Fix by replacing the call to map() in the
Search.display_results() method with a list comprehension.

Signed-off-by: Daniel Harding <dharding@living180.net>