gitcmds: handle submodules on empty branch
The diff_index() function has to be careful to handle the case
of a branch without any commits (e.g. immediately after a git init or a
git checkout --orphan).  In that case there is no HEAD commit to compare
against, which causes the git diff-index --cached HEAD call to fail.  In
that case the diff_index() function was simply returning the list of
tracked files.  However, if there were any submodules, they would not be
recognized as such and would be treated as normal files.  To address
this, if the initial git diff-index call fails, re-run it against the
SHA1 of the empty tree instead of against HEAD, and then continue with
the normal code path.

Signed-off-by: Daniel Harding <dharding@living180.net>