gitcmds: remove unneeded code in diff_worktree()
The diff_worktree() function had code to handle an error from the
git diff-files call, with a comment stating that it was for git init
(e.g. no commits yet).  However, that case does not actually cause an
error, because git diff-files compares the working tree with the index,
and the index always exists, even if there are no commits yet.  So,
remove the code for handling an error from git diff-files.

Signed-off-by: Daniel Harding <dharding@living180.net>