widgets/status: use icon for untracked files
git-cola has an icon for untracked files, but it wasn't being used,
because the StatusTreeWidget was not providing the correct argument to
qtutils.create_treeitem().  Change StatusTreeWidget.set_untracked() to
pass untracked=True to StatusTreeWidget._set_subtree(), which will pass
that value on to qtutils.create_treeitem(), causing the desired icon to
be used.

Signed-off-by: Daniel Harding <dharding@living180.net>