gitcmds: include submodules in path lists
Change diff_index() and diff_worktree() to include submodules in the
returned path lists, rather than adding them to the path lists in
worktree_state().  This simplifies the code and also avoids a bug where
an unmerged submodule would be displayed in the Staged section of the
Status widget instead of the Unmerged section.

Signed-off-by: Daniel Harding <dharding@living180.net>