widgets/status: fix unstaged deleted diff
2cbd38e2edca6ae6aa732b7afb809fb4b2a7dbf0 broke displaying the diff for
an unstaged deleted file.  In the call to cmds.do(cmds.Diff, ...),
item.deleted was being passed as a positional argument, which meant it
was actually being assigned to the cached argument, rather than the
deleted argument, resulting in an empty diff.  Change to pass as a
keyword argument to avoid this.

Signed-off-by: Daniel Harding <dharding@living180.net>