interaction: add an Interaction.safe_log() method
This method does the same thing as Interaction.log(), except that it can
be called from threads other than the main thread.  For the default
interaction class, safe_log() is exactly the same as the log() method.
However, when Interaction gets patched so that the log() method actually
calls the LogWidget.log() method, its safe_log() method also gets
patched to call the new LogWidget.safe_log() method.  The
LogWidget.safe_log() method invokes the LogWidget.log() method through a
signal, so that the log() method actually runs on the main thread.

Signed-off-by: Daniel Harding <dharding@living180.net>