diff: remember the text cursor's position when updating diffs
Previously, the text cursor would often jump to the top of the diff
after staging, which makes it difficult to use the keyboard arrows to
select and stage lines.  Using the arrows to move the cursor would cause
the display to jump to the top of the file, which counter-acts the
careful scrollbar placement which is done when updating the diff text.

Do a better job of remember the text cursor's position within the diff
text.  This makes it easier to interactively use the keyboard arrows to
select and stage lines by arranging for the cursor to be nearby its
original location after staging.

Signed-off-by: David Aguilar <davvid@gmail.com>