app: implement refresh-on-focus
Implement the ability to refresh git cola whenever it gains focus.  This
functionality is disabled by default, but can be enabled by setting the
cola.refreshonfocus git config setting to true.  It is not enabled by
default because it can result in a pause whenever switching to git cola
from another application (especially noticeable on Windows).

Requested-by: David LeGare <excaliburHisSheath@gmail.com>
Signed-off-by: Daniel Harding <dharding@living180.net>