fsmonitor: track changes within the git directory
To keep git-cola's status up-to-date when changes are made to the
repository from outside git-cola (e.g. by invoking git from the command
line), monitor the git repository and refresh when it changes.

When using inotify, monitor the git directory itself (to refresh on
changes to HEAD or index) along the refs subdirectory and all its
subdirectories.

On Windows, open a separate handle for the git directory in addition to
the handle for the worktree.  To make the code cleaner, this commit
moves the details of using the Windows ReadDirectoryChanges API into a
new class, _Win32Watch.

Signed-off-by: Daniel Harding <dharding@living180.net>