gitcmds: force option "no-abbrev-commit" when running "git log"
I have set "log.abbrevCommit = true" in my ".gitconfig" file.

In that case the "More..." items in the "Load previous Commit Message" and
"Fixup previous Commit" sub menus wont work.

The reason is the regular expression REV_LIST_REGEX. It expects the SHA1 to be
exactly 40 characters long in the git log output.

Signed-off-by: Wolfgang Ocker <weo@reccoware.de>