rebase: use the same cli syntax as `git rebase`
Make `git cola rebase` behave the same as `git rebase` on the
command line.  We now honor [--onto <newbase>] [upstream]
[branch], and the rebase will now default to @{upstream} when
it is not specified.

Internally, we let `git rebase`'s defaults kick in by simply not
specifying anything when <upstream> is omitted.  This
effectively makes the first argument to `git cola rebase`, which
was previously required, optional.

Closes #416
Suggested-by: Patrick Browne <patrick.browne@amadeus.com>
Signed-off-by: David Aguilar <davvid@gmail.com>