xbase: optimize decorate() calls
Only decorate items that moved during an internal drag by hooking into
the items_moved signal from the base class.

When performing an internal move via the hotkeys, we only need to
decorate the item that moved, unless when we've moved to the zeroeth
position, at which point all items must be re-decorated.

Signed-off-by: David Aguilar <davvid@gmail.com>