widgets.completion: detect when a line is cleared
Emit a signal when the text field cleared.  This will allow widgets to
perform special actions, such as updating their filters.

Hide the completion popup when the line is clear and backspace is
pressed again.  This makes it easy to hide the completion dialog (and
avoid completions) by pressing backspace until it disappears.

Signed-off-by: David Aguilar <davvid@gmail.com>