status: update the path filter when a line is cleared
Hook into the cleared() signal so that the path filter is updated when a
line is cleared.  Also hook into return() and editingFinished() so that
we catch all possible ways for the filter text to be updated.

Remember the previous filter value so that we can skip updating the
paths filter if updating it would produce the same result.

Signed-off-by: David Aguilar <davvid@gmail.com>