Windows packaging using Pynsist
This is still very rough. In particular, it should be smarter about
finding the git executable - it currently assumes that it's installed in
Program Files (the default location).

To build:

1. Download PyQt4
(http://sourceforge.net/projects/pyqt/files/PyQt4/PyQt-4.11.3/PyQt4-4.11.3
-gpl-Py3.4-Qt4.8.6-x32.exe ) and unpack files from Lib/site-packages
into pynsist_pkgs/ in the root of the git-cola repo. Then add files from
   $_OUTDIR/ into pynsist_pkgs/PyQt4. I hope to automate this step in
the future.
2. rm share/doc/git-cola/html - this is a symlink, and copying that into
the build directory currently fails.
3. With pynsist and nsis installed, run 'pynsist windows-installer.cfg'.
The installer will be produced in build/nsis/