doc: eliminate the share/doc/git-cola/html symlink
We've traditionally relied on a symlink in the worktree to allow
finding git-cola's documentation when running from within a
git-cola sandbox.

We can make this simpler by explicitly falling back to the
_build/html/ path rather than depending on the worktree to
contain a symlink.

This also makes things better for portability, as was pointed
out in #421.

Update resources.html_docs() to reflect the new scheme.

Related-to: #421
Noticed-by: Thomas Kluyver <takowl@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>