models.main: improve commit message auto-load behavior
Move all of the commit message auto-load logic into the model.

When automatically loading the commit message from e.g. .git/MERGE_MSG,
remember the current message so that we can restore it later.

When we detect that the message paths no longer exist, and the commit
message has not been edited by the user, then reset the message back to
the original message.

Closes #431
Suggested-by: Dave Thomas <dave@monstergfx.com>
Signed-off-by: David Aguilar <davvid@gmail.com>