dag: handle Git v2.4 behavior change for `git log --decorate`
Git v2.4 (currently a release candidate) introduces a change in
the output from `git log --decorate`, which we parse in the DAG:

Git 2.4 Release Notes (draft)
=============================

Backward compatibility warning(s)
---------------------------------

This release has a few changes in the user-visible output from
Porcelain commands. These are not meant to be parsed by scripts, but
the users still may want to be aware of the changes:

 * Output from "git log --decorate" (and "%d" format specifier used in
   the userformat "--format=<string>" parameter "git log" family of
   command takes) used to list "HEAD" just like other tips of branch
   names, separated with a comma in between.  E.g.

     $ git log --decorate -1 master
     commit bdb0f6788fa5e3cacc4315e9ff318a27b2676ff4 (HEAD, master)
     ...

   This release updates the output slightly when HEAD refers to the tip
   of a branch whose name is also shown in the output.  The above is
   shown as:

     $ git log --decorate -1 master
     commit bdb0f6788fa5e3cacc4315e9ff318a27b2676ff4 (HEAD -> master)
     ...

Add a add_label() function and handle the format change there
so that the behavior remains consistent before and after Git v2.4.

C.f. http://thread.gmane.org/gmane.linux.kernel/1931234

Signed-off-by: David Aguilar <davvid@gmail.com>